+++ CHANGELOG.md
@@ -8,5 +8,16 @@ Removed: Backward-incompatible removals.
 Fixed: Bug fixes.
 Security: Security patches (critical to highlight). 
 
+## [0.2.1] - 2025-05-19
+### Added
+- example_account_info.py file to show examples of using account_info.py module
+- added def get_full_model_detail_dict method to VeniceModels class
+
+### Removed
+- removed application specific methods to show VeniceModel info in specific format.  Those should be in programs.
+-- get_model_detail_dict
+-- get_model_detail
+
+
 ## [0.2.0] - 2025-05-16
 - Initial stable release with CHANGELOG